Restructuring, stage for es6 base without script-loader (which is deprecated anyway)
